fix: rearrange Home.js with useHomeFetch
